requirements.txt
Flask==3.1.1
python-dotenv==1.1.1
pandas==2.1.1
requests==2.32.4
Flask-Limiter==3.3.1
matplotlib==3.8.0
weasyprint>=66.0
gunicorn==20.1.0
black
flake8
pytest